Extract NewItemForm, StoreHeader, & StoreNotes components from Store
Motivation: smaller components are easier to maintain. A bunch of
different concerns were previously all being mashed together in the big
Store component.
